requirements.txt
Django==1.10.3
django-filter==1.0.0
djangorestframework==3.5.3
get==0.0.20
gunicorn==19.6.0
Markdown==2.6.7
numpy==1.11.2
pandas==0.19.1
post==0.0.12
public==0.0.37
Pygments==2.1.3
python-dateutil==2.6.0
pytz==2016.7
query-string==0.0.11
redis==2.10.5
request==0.0.12
scikit-learn==0.18.1
setupfiles==0.0.13
six==1.10.0
uWSGI==2.0.14
